requirements-test.txt
Fabric==1.8.0
PyHamcrest==1.8.0
nose==1.3.0
nose-testconfig==0.9
requests==2.0.1
splinter==0.5.4
